refactor: wal module structuring
