+++ src/wal/wal.rs
@@ -1,132 +1,21 @@
 use std::io::{self, BufReader, Read};
 use std::path::Path;
-use std::sync::Arc;
 use std::{fs::File, io::Write, path::PathBuf};
 
 use bytes::{BufMut, Bytes, BytesMut};
 use hashbrown::HashMap;
-use parking_lot::Mutex;
 use prost::Message;
 use tracing::{debug, error, info};
 use walkdir::{DirEntry, WalkDir};
 
+use super::DEFAULT_MAX_SEGMENT_SIZE;
+use super::MAX_SEGMENT_BUFFER_SIZE;
+use super::WAL_DEFAULT_ID;
+use super::WAL_EXTENSION;
 use crate::data_types::wal::segment_entry::Entry as EntryProto;
 use crate::data_types::wal::SegmentEntry as SegmentEntryProto;
 use crate::data_types::{BigPipeValue, RetentionPolicy, WalOperation};
 use crate::ServerMessage;
-
-pub(crate) const DEFAULT_MAX_SEGMENT_SIZE: usize = 16777216; // 16 MiB
-const MAX_SEGMENT_BUFFER_SIZE: u16 = 8192; // 8 KiB
-
-const WAL_EXTENSION: &str = "-bp.wal";
-pub(crate) const WAL_DEFAULT_ID: u64 = 0;
-
-/// A write-ahead log implementation which will handle
-/// multiple underlying [`Wal`]s at once, keyed by
-/// namespace.
-#[derive(Debug)]
-pub struct MultiWal {
-    namespaces: Arc<Mutex<HashMap<String, Wal>>>,
-    root_directory: PathBuf,
-    max_segment_size: usize,
-}
-
-impl MultiWal {
-    /// Create a new instance of [`MultiWal`].
-    pub fn new(root_directory: PathBuf, max_segment_size: Option<usize>) -> Self {
-        Self {
-            namespaces: Arc::new(Mutex::new(HashMap::with_capacity(100))),
-            root_directory,
-            max_segment_size: max_segment_size.unwrap_or(DEFAULT_MAX_SEGMENT_SIZE),
-        }
-    }
-
-    /// Create the WAL directory structure that will be used for the underlying
-    /// [`Wal`] for a particular key, returning the associated [`PathBuf`].
-    fn create_wal_directory(&self, namespace: &str) -> PathBuf {
-        let wal_dir = PathBuf::from(format!("{}/{namespace}", self.root_directory.display()));
-        std::fs::create_dir(&wal_dir).unwrap();
-        wal_dir
-    }
-
-    #[allow(dead_code)]
-    /// Flush the [`Wal`] of a particular namespace.
-    pub(crate) fn flush(&self, namespace: &str) -> Result<(), Box<dyn std::error::Error>> {
-        self.namespaces.lock().get_mut(namespace).unwrap().flush()
-    }
-
-    #[allow(dead_code)]
-    /// Flush the [`Wal`] of all namespaces.
-    pub(crate) fn flush_all(&self) -> Result<(), Box<dyn std::error::Error>> {
-        self.namespaces
-            .lock()
-            .iter_mut()
-            .try_for_each(|(_, wal)| wal.flush())
-    }
-
-    fn write_under_lock(&self, namespace: &str, op: &WalOperation) {
-        self.namespaces
-            .lock()
-            .entry(namespace.to_string())
-            .and_modify(|wal| {
-                wal.write(op.clone()).unwrap();
-            })
-            .or_insert_with(|| {
-                let wal_dir = self.create_wal_directory(namespace);
-                let mut wal =
-                    Wal::try_new(WAL_DEFAULT_ID, wal_dir, Some(self.max_segment_size)).unwrap();
-                wal.write(op.clone()).unwrap(); // Ensure that the inbound op is not lost
-                wal
-            });
-    }
-
-    /// Write a [`WalOperation`] to the respective [`Wal`]. The key within
-    /// the operation is used as the namespace.
-    pub fn write(&self, op: WalOperation) -> Result<(), Box<dyn std::error::Error>> {
-        match &op {
-            WalOperation::Message(msg) => {
-                self.write_under_lock(&msg.key, &op);
-                Ok(())
-            }
-            WalOperation::Namespace(namespace) => {
-                self.write_under_lock(&namespace.key, &op);
-                Ok(())
-            }
-        }
-    }
-
-    /// Replay all [`Wal`] files that are found from the given directory.
-    ///
-    /// This will walk the given directory, it is a expectation that
-    /// sub-directories are keys for a namespace and will contain
-    /// individual segment files.
-    ///
-    /// For example:
-    ///
-    /// /tmp/example-multi-wal
-    /// ├── bar <-- namespace 'bar'
-    /// │   └── 0-bp.wal
-    /// └── foo <-- namespace 'foo'
-    ///     ├── 0-bp.wal
-    ///     └── 1-bp.wal
-    pub fn replay<P: AsRef<Path>>(directory: P) -> HashMap<String, BigPipeValue> {
-        let mut multi = HashMap::new();
-        for entry in WalkDir::new(directory)
-            .max_depth(1) // current directory only
-            .into_iter()
-        {
-            let entry = entry.unwrap();
-            // Hack to skip over the initially passed
-            // directory, as this counts as an entry.
-            // Find a better way to do this.
-            if entry.depth() == 1 {
-                let (_, inner) = Wal::replay(entry.path()).expect("should exist");
-                multi.extend(inner);
-            }
-        }
-        multi
-    }
-}
 
 /// A write-ahead log (WAL) implementation.
 ///
@@ -305,43 +194,6 @@ impl Wal {
     }
 }
 
-// Adapted from <https://docs.rs/wyre/latest/src/wyre/lib.rs.html#53-70>
-fn read_varint<R: Read>(reader: &mut R) -> io::Result<Option<usize>> {
-    let mut buf = [0u8; 1];
-    let mut shift = 0;
-    let mut result: usize = 0;
-
-    for _ in 0..10 {
-        if reader.read(&mut buf)? == 0 {
-            return Ok(None); // EOF
-        }
-        let byte = buf[0];
-        result |= ((byte & 0x7F) as usize) << shift;
-        if byte & 0x80 == 0 {
-            return Ok(Some(result));
-        }
-        shift += 7;
-    }
-
-    Err(io::Error::new(
-        io::ErrorKind::InvalidData,
-        "varint too long",
-    ))
-}
-
-fn parse_segment_id(entry: &DirEntry) -> u64 {
-    debug!(?entry);
-    entry
-        .path()
-        .file_name()
-        .unwrap()
-        .to_string_lossy()
-        .strip_suffix(WAL_EXTENSION)
-        .expect("Only WAL files are returned by the filter")
-        .parse::<u64>()
-        .unwrap()
-}
-
 #[derive(Debug)]
 struct Segment {
     filepath: PathBuf,
@@ -398,6 +250,43 @@ impl Segment {
     fn path(&self) -> &Path {
         &self.filepath
     }
+}
+
+// Adapted from <https://docs.rs/wyre/latest/src/wyre/lib.rs.html#53-70>
+fn read_varint<R: Read>(reader: &mut R) -> io::Result<Option<usize>> {
+    let mut buf = [0u8; 1];
+    let mut shift = 0;
+    let mut result: usize = 0;
+
+    for _ in 0..10 {
+        if reader.read(&mut buf)? == 0 {
+            return Ok(None); // EOF
+        }
+        let byte = buf[0];
+        result |= ((byte & 0x7F) as usize) << shift;
+        if byte & 0x80 == 0 {
+            return Ok(Some(result));
+        }
+        shift += 7;
+    }
+
+    Err(io::Error::new(
+        io::ErrorKind::InvalidData,
+        "varint too long",
+    ))
+}
+
+fn parse_segment_id(entry: &DirEntry) -> u64 {
+    debug!(?entry);
+    entry
+        .path()
+        .file_name()
+        .unwrap()
+        .to_string_lossy()
+        .strip_suffix(WAL_EXTENSION)
+        .expect("Only WAL files are returned by the filter")
+        .parse::<u64>()
+        .unwrap()
 }
 
 #[cfg(test)]
@@ -573,65 +462,5 @@ mod test {
 
         let wal = Wal::try_new(10, dir.path().to_path_buf(), None).unwrap();
         assert_eq!(wal.closed_segments().len(), 10); // 0-base index, 0..=9 is 10 closed segments
-    }
-
-    #[test]
-    fn multi_directory_structure() {
-        let dir = TempDir::new().unwrap();
-        let multi = MultiWal::new(dir.path().to_path_buf(), None);
-
-        let wal_dir = multi.create_wal_directory("my_new_namespace");
-        assert_eq!(
-            wal_dir.to_string_lossy(),
-            format!("{}/my_new_namespace", dir.path().display())
-        );
-        assert!(std::fs::exists(&wal_dir).unwrap());
-    }
-
-    #[test]
-    fn multi_wal_ops() {
-        let dir = TempDir::new().unwrap();
-        let multi = MultiWal::new(dir.path().to_path_buf(), None);
-
-        multi.write(WalOperation::test_message(10)).unwrap();
-        multi.flush("hello").unwrap();
-
-        drop(multi);
-
-        let previous_dir = format!("{}/hello", dir.path().display());
-        let (_, inner) = Wal::replay(&previous_dir).unwrap();
-        assert_eq!(inner.get("hello").unwrap().get_range(0).len(), 1);
-        assert_eq!(
-            inner.get("hello").unwrap().get_range(0),
-            vec![ServerMessage::new("hello".to_string(), "world".into(), 10)]
-        );
-    }
-
-    #[test]
-    fn multi_replay() {
-        let dir = TempDir::new().unwrap();
-        let multi = MultiWal::new(dir.path().to_path_buf(), None);
-
-        multi
-            .write(WalOperation::test_message(10).with_key("foo"))
-            .unwrap();
-        multi
-            .write(WalOperation::test_message(20).with_key("bar"))
-            .unwrap();
-
-        multi.flush_all().unwrap();
-
-        drop(multi);
-
-        let multi = MultiWal::replay(dir.path());
-        assert_eq!(multi.keys().len(), 2);
-        assert_eq!(
-            multi.get("foo").unwrap().get_range(0),
-            vec![ServerMessage::new("foo".to_string(), "world".into(), 10)]
-        );
-        assert_eq!(
-            multi.get("bar").unwrap().get_range(0),
-            vec![ServerMessage::new("bar".to_string(), "world".into(), 20)]
-        );
     }
 }
